// client/src/contracts/types/ERC20TokensSwap.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from "bn.js";
import { ContractOptions } from "web3-eth-contract";
import { EventLog } from "web3-core";
import { EventEmitter } from "events";
import {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export type Rate = ContractEventLog<{
  scaledPrice: string;
  timeStamp: string;
  0: string;
  1: string;
}>;
export type SwapRateInfo = ContractEventLog<{
  exchangeType: string;
  sellingAmount: string;
  buyingAmount: string;
  0: string;
  1: string;
  2: string;
}>;
export type SwapTransferInfo = ContractEventLog<{
  beneficiary: string;
  amountSent: string;
  amountReceived: string;
  0: string;
  1: string;
  2: string;
}>;

export interface ERC20TokensSwap extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): ERC20TokensSwap;
  clone(): ERC20TokensSwap;
  methods: {
    baseToken(): NonPayableTransactionObject<string>;

    getAskPrice(
      _baseTokenAmount: number | string | BN
    ): NonPayableTransactionObject<string>;

    getAvailableBaseTokenLiquidity(): NonPayableTransactionObject<string>;

    getAvailableQuoteTokenLiquidity(): NonPayableTransactionObject<string>;

    getBidPrice(
      _quoteTokenAmount: number | string | BN
    ): NonPayableTransactionObject<string>;

    getScaledRate(
      _scalingDecimal: number | string | BN
    ): NonPayableTransactionObject<string>;

    pairName(): NonPayableTransactionObject<string>;

    quoteToken(): NonPayableTransactionObject<string>;

    swapBaseForQuoteToken(
      _baseTokenAmount: number | string | BN
    ): PayableTransactionObject<void>;

    swapQuoteForBaseToken(
      _quoteTokenAmount: number | string | BN
    ): PayableTransactionObject<void>;
  };
  events: {
    Rate(cb?: Callback<Rate>): EventEmitter;
    Rate(options?: EventOptions, cb?: Callback<Rate>): EventEmitter;

    SwapRateInfo(cb?: Callback<SwapRateInfo>): EventEmitter;
    SwapRateInfo(
      options?: EventOptions,
      cb?: Callback<SwapRateInfo>
    ): EventEmitter;

    SwapTransferInfo(cb?: Callback<SwapTransferInfo>): EventEmitter;
    SwapTransferInfo(
      options?: EventOptions,
      cb?: Callback<SwapTransferInfo>
    ): EventEmitter;

    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };

  once(event: "Rate", cb: Callback<Rate>): void;
  once(event: "Rate", options: EventOptions, cb: Callback<Rate>): void;

  once(event: "SwapRateInfo", cb: Callback<SwapRateInfo>): void;
  once(
    event: "SwapRateInfo",
    options: EventOptions,
    cb: Callback<SwapRateInfo>
  ): void;

  once(event: "SwapTransferInfo", cb: Callback<SwapTransferInfo>): void;
  once(
    event: "SwapTransferInfo",
    options: EventOptions,
    cb: Callback<SwapTransferInfo>
  ): void;
}