// client/src/contracts/types/SwapContractFactory.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import type BN from "bn.js";
import type { ContractOptions } from "web3-eth-contract";
import type { EventLog } from "web3-core";
import type { EventEmitter } from "events";
import type {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export type DeployedSwapContract = ContractEventLog<{
  swapContractAddress: string;
  baseTokenAddress: string;
  quoteTokenAddress: string;
  0: string;
  1: string;
  2: string;
}>;
export type OwnershipTransferred = ContractEventLog<{
  previousOwner: string;
  newOwner: string;
  0: string;
  1: string;
}>;

export interface SwapContractFactory extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): SwapContractFactory;
  clone(): SwapContractFactory;
  methods: {
    deployNewSwapContract(
      _baseToken: string,
      _quoteToken: string,
      _exchangeRate: string
    ): NonPayableTransactionObject<void>;

    deployedSwapContractsRegistry(arg0: string): NonPayableTransactionObject<{
      swapContractAddress: string;
      quoteTokenAddress: string;
      baseTokenAddress: string;
      deployed: boolean;
      0: string;
      1: string;
      2: string;
      3: boolean;
    }>;

    getAllSwapPairs(): NonPayableTransactionObject<string[]>;

    owner(): NonPayableTransactionObject<string>;

    renounceOwnership(): NonPayableTransactionObject<void>;

    swapPairs(arg0: number | string | BN): NonPayableTransactionObject<string>;

    transferOwnership(newOwner: string): NonPayableTransactionObject<void>;
  };
  events: {
    DeployedSwapContract(cb?: Callback<DeployedSwapContract>): EventEmitter;
    DeployedSwapContract(
      options?: EventOptions,
      cb?: Callback<DeployedSwapContract>
    ): EventEmitter;

    OwnershipTransferred(cb?: Callback<OwnershipTransferred>): EventEmitter;
    OwnershipTransferred(
      options?: EventOptions,
      cb?: Callback<OwnershipTransferred>
    ): EventEmitter;

    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };

  once(event: "DeployedSwapContract", cb: Callback<DeployedSwapContract>): void;
  once(
    event: "DeployedSwapContract",
    options: EventOptions,
    cb: Callback<DeployedSwapContract>
  ): void;

  once(event: "OwnershipTransferred", cb: Callback<OwnershipTransferred>): void;
  once(
    event: "OwnershipTransferred",
    options: EventOptions,
    cb: Callback<OwnershipTransferred>
  ): void;
}